refactor(smtp_server): extract 3 helpers purs (recipient, tls, auth)
Split de src/bin/smtp_server.rs (1129 LOC) vers 3 sous-modules:
- smtp_server_dir/recipient.rs (44 LOC) — recipient_domain, is_local_recipient, recipient_local_part
- smtp_server_dir/tls.rs (15 LOC) — load_certs, load_key
- smtp_server_dir/auth.rs (17 LOC) — check_credentials

Extractions à zéro risque : fonctions pures sans dépendance métier.
smtp_server.rs: 1129 → 1061 LOC (−6%).

Les handlers plus gros (handle_tls_client, handle_plain_client, process_command,
apply_parsed_header) ont beaucoup d'interdépendances avec CustomEmail/MailServer
et nécessitent un refactor plus profond — à faire dans un PR ultérieur.
